Use AtomicIntegers and adjust bar dimensions

+++ src/ie/dcu/Car.java
@@ -15,7 +15,7 @@ public class Car {
         }
     }
 
-    public double getSpace() {
+    public int getSpace() {
         return space;
     }
